Convert test_prepare_proposal_vext_normal_op() to a mempool test
